[Fix][Tests] Add OOB negative test; isolate overrides/config via system properties; guard asset copying in tests

+++ apps/src/main/scala/com/crib/bills/dom6maps/services/mapeditor/MapEditorCopier.scala
@@ -45,9 +45,12 @@ class MapEditorCopierImpl[Sequencer[_]: Async: Files] extends MapEditorCopier[Se
             else if maybeMain.contains(p) || maybeCave.contains(p) then Async[Sequencer].unit
             else
               val parent = target.parent.getOrElse(dest)
-              Files[Sequencer].createDirectories(parent) >>
-                Files[Sequencer].deleteIfExists(target) >>
-                Files[Sequencer].copy(p, target)
+              val skipAssets = sys.props.get("dom6.skipAssetCopy").contains("true")
+              if skipAssets then Async[Sequencer].unit
+              else
+                Files[Sequencer].createDirectories(parent) >>
+                  Files[Sequencer].deleteIfExists(target) >>
+                  Files[Sequencer].copy(p, target)
           }
       result <- maybeMain match
         case Some(mapPath) =>
